backend/requirements.txt
fastapi
uvicorn
pydantic
openai
pdfplumber
faiss-cpu
numpy
